% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katpackage.R
\name{OR_95CI}
\alias{OR_95CI}
\title{OR Function}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{coefficient from regression model}

\item{se}{standard error}

\item{siglevel}{significance level}

\item{roundto}{number of decimals}
}
\value{
OR, 95% CI
}
\description{
Using coef and SE, specify significance level and number of decimal places to get OR and 95% CI
}
\examples{
logregr <- glm(y ~ x1 + x2, data=toydata, family = "binomial")
regsummary <- summary(logregr)
coef <- regsummary$coefficients[,1]
se <- regsummary$coefficients[,2]
OR_95CI(coef,se,0.05,2)
}
\author{
Katherine Lu
}
