% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{y}{binary outcome}
  \item{x1}{continuous predictor}
  \item{x2}{binary predictor}
}
}
\usage{
toydata
}
\description{
Example data
}
\keyword{datasets}
